requirements_good_ubuntu__django314_python375.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.4
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-meta==2.0.0
django-nested-admin==3.2.3
django-ratelimit==3.0.1
django-request==1.5.6
django-sass-processor==0.8.2
djangorestframework==3.12.2
Faker==5.0.1
idna==2.8
importlib-metadata==3.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
libsass==0.20.1
lorem==0.1.1
Markdown==3.3.3
mccabe==0.6.1
nodeenv==1.5.0
oauthlib==3.1.0
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylibmc==1.6.1
pylint==2.6.0
pylint-django==2.4.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-gettext==4.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2019.1
rcssmin==1.0.6
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2
sqlparse==0.3.0
text-unidecode==1.3
tldextract==2.2.2
toml==0.10.2
typed-ast==1.4.2
urllib3==1.25.6
wrapt==1.12.1
zipp==3.4.0
